Added scrollEvents parameter
Added a scrollEvents parameter that can be set by the user.  If set to
false, plugin will ignore scroll events and will not animate when they
are detected such as when the scroll wheel is moved or someone scrolls
on a mobile device.  Is useful when you want to only scroll using the
moveUp and moveDown methods.